fix: eliminate unnecessary allocations, duplicated logic, and diff parsing bugs
It turns out the codebase had a collection of issues that were each
individually minor but collectively *not great*.

parse_diff_output() was clone()-ing vectors then immediately
clear()-ing them, which is the kind of thing that makes you wonder
if anyone ever profiled this. Replaced with std::mem::take(). The
metadata skip list was also missing about nine git diff metadata
prefixes (rename from, similarity index, Binary files, etc.), so
renamed and binary files would leak metadata lines into the diff
content as spurious context lines. While at it, the old
starts_with("new") was hilariously over-broad — it matched any
diff content starting with "new". Fixed to match the actual git
metadata prefixes.

The scroll position was stored as u16, silently capping at 65535
lines. For a diff viewer. The accept/reject handlers were
copy-pasted with one enum variant changed. The n/p file navigation
was duplicated with mirrored loop directions. The upstream branch
lookup was hand-rolled in two places. Context lines in the syntax
highlighter carried a stale leading space from the git diff format,
misaligning them with changed lines and confusing syntect's parser
for indentation-sensitive languages.

Let's just fix all of it: u16 -> usize for scroll, extract helpers
for duplicated logic, extract get_upstream_branch() into diff.rs,
strip the diff-format leading space before highlighting, pass
--no-color to git diff so we don't rely on regex stripping, and
exit early with "No changes." instead of launching a full TUI for
an empty diff. Added 16 new tests covering the previously untested
paths.

+++ src/ui/types.rs
@@ -33,7 +33,7 @@ pub struct App<'a> {
     pub right_label: &'a str,
     pub current_file_idx: usize,
     pub file_names: Vec<String>,
-    pub scroll_positions: HashMap<String, u16>,
+    pub scroll_positions: HashMap<String, usize>,
     pub focused_pane: Pane,
     pub view_mode: ViewMode,
     pub app_mode: AppMode,
